Refactor datagen so that addons have a clear API to follow.
