+++ src/main/java/com/simibubi/create/api/data/recipe/CuttingRecipeGen.java
@@ -1,0 +1,30 @@
+package com.simibubi.create.api.data.recipe;
+
+import com.simibubi.create.AllRecipeTypes;
+
+import net.minecraft.data.PackOutput;
+import net.minecraft.world.level.block.Block;
+
+@SuppressWarnings("unused")
+public class CuttingRecipeGen extends ProcessingRecipeGen {
+
+	protected GeneratedRecipe stripAndMakePlanks(Block wood, Block stripped, Block planks) {
+		return stripAndMakePlanks(wood, stripped, planks, 6);
+	}
+
+	protected GeneratedRecipe stripAndMakePlanks(Block wood, Block stripped, Block planks, int planksAmount) {
+		create(() -> wood, b -> b.duration(50)
+			.output(stripped));
+		return create(() -> stripped, b -> b.duration(50)
+			.output(planks, planksAmount));
+	}
+
+	public CuttingRecipeGen(PackOutput output, String defaultNamespace) {
+		super(output, defaultNamespace);
+	}
+
+	@Override
+	protected AllRecipeTypes getRecipeType() {
+		return AllRecipeTypes.CUTTING;
+	}
+}
