+++ src/main/java/com/simibubi/create/api/data/recipe/EmptyingRecipeGen.java
@@ -1,0 +1,19 @@
+package com.simibubi.create.api.data.recipe;
+
+import com.simibubi.create.AllRecipeTypes;
+
+
+import net.minecraft.data.PackOutput;
+
+public class EmptyingRecipeGen extends ProcessingRecipeGen {
+
+	public EmptyingRecipeGen(PackOutput output, String defaultNamespace) {
+		super(output, defaultNamespace);
+	}
+
+	@Override
+	protected AllRecipeTypes getRecipeType() {
+		return AllRecipeTypes.EMPTYING;
+	}
+
+}
