+++ src/main/java/com/simibubi/create/api/data/recipe/FillingRecipeGen.java
@@ -1,0 +1,18 @@
+package com.simibubi.create.api.data.recipe;
+
+import com.simibubi.create.AllRecipeTypes;
+
+import net.minecraft.data.PackOutput;
+
+public abstract class FillingRecipeGen extends ProcessingRecipeGen {
+
+	public FillingRecipeGen(PackOutput output, String defaultNamespace) {
+		super(output, defaultNamespace);
+	}
+
+	@Override
+	protected AllRecipeTypes getRecipeType() {
+		return AllRecipeTypes.FILLING;
+	}
+
+}
