+++ src/main/java/com/simibubi/create/api/data/recipe/MixingRecipeGen.java
@@ -1,0 +1,18 @@
+package com.simibubi.create.api.data.recipe;
+
+import com.simibubi.create.AllRecipeTypes;
+
+import net.minecraft.data.PackOutput;
+
+public class MixingRecipeGen extends ProcessingRecipeGen {
+
+	public MixingRecipeGen(PackOutput output, String defaultNamespace) {
+		super(output, defaultNamespace);
+	}
+
+	@Override
+	protected AllRecipeTypes getRecipeType() {
+		return AllRecipeTypes.MIXING;
+	}
+
+}
