+++ src/main/java/com/simibubi/create/api/data/recipe/PolishingRecipeGen.java
@@ -1,20 +1,15 @@
-package com.simibubi.create.foundation.data.recipe;
+package com.simibubi.create.api.data.recipe;
 
 import com.simibubi.create.AllItems;
 import com.simibubi.create.AllRecipeTypes;
 
+import com.simibubi.create.foundation.data.recipe.CreateRecipeProvider;
 import net.minecraft.data.PackOutput;
 
 public class PolishingRecipeGen extends ProcessingRecipeGen {
 
-	GeneratedRecipe
-
-	ROSE_QUARTZ = create(AllItems.ROSE_QUARTZ::get, b -> b.output(AllItems.POLISHED_ROSE_QUARTZ.get()))
-
-	;
-
-	public PolishingRecipeGen(PackOutput output) {
-		super(output);
+	public PolishingRecipeGen(PackOutput output, String defaultNamespace) {
+		super(output, defaultNamespace);
 	}
 
 	@Override
