+++ src/main/java/com/simibubi/create/api/data/recipe/PressingRecipeGen.java
@@ -1,0 +1,18 @@
+package com.simibubi.create.api.data.recipe;
+
+import com.simibubi.create.AllRecipeTypes;
+
+import net.minecraft.data.PackOutput;
+
+public class PressingRecipeGen extends ProcessingRecipeGen {
+
+	public PressingRecipeGen(PackOutput output, String defaultNamespace) {
+		super(output, defaultNamespace);
+	}
+
+	@Override
+	protected AllRecipeTypes getRecipeType() {
+		return AllRecipeTypes.PRESSING;
+	}
+
+}
