+++ src/main/java/com/simibubi/create/api/data/recipe/SequencedAssemblyRecipeGen.java
@@ -1,0 +1,44 @@
+package com.simibubi.create.api.data.recipe;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.function.Consumer;
+import java.util.function.UnaryOperator;
+import java.util.stream.Stream;
+
+import com.simibubi.create.AllBlocks;
+import com.simibubi.create.AllItems;
+import com.simibubi.create.AllTags.AllItemTags;
+import com.simibubi.create.Create;
+import com.simibubi.create.content.fluids.transfer.FillingRecipe;
+import com.simibubi.create.content.kinetics.deployer.DeployerApplicationRecipe;
+import com.simibubi.create.content.kinetics.press.PressingRecipe;
+import com.simibubi.create.content.processing.sequenced.SequencedAssemblyRecipeBuilder;
+
+import com.simibubi.create.foundation.data.recipe.CreateRecipeProvider;
+import net.minecraft.data.PackOutput;
+import net.minecraft.data.recipes.FinishedRecipe;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraft.world.level.material.Fluids;
+
+public class SequencedAssemblyRecipeGen extends BaseRecipeProvider {
+
+	public SequencedAssemblyRecipeGen(PackOutput output, String defaultNamespace) {
+		super(output, defaultNamespace);
+	}
+
+	@Override
+	public String getName() {
+		return modid + "'s sequenced assembly recipes";
+	}
+
+	protected GeneratedRecipe create(String name, UnaryOperator<SequencedAssemblyRecipeBuilder> transform) {
+		GeneratedRecipe generatedRecipe =
+			c -> transform.apply(new SequencedAssemblyRecipeBuilder(asResource(name)))
+				.build(c);
+		all.add(generatedRecipe);
+		return generatedRecipe;
+	}
+}
