+++ src/main/java/com/simibubi/create/foundation/data/recipe/CreatePolishingRecipeGen.java
@@ -1,0 +1,23 @@
+package com.simibubi.create.foundation.data.recipe;
+
+import com.simibubi.create.AllItems;
+import com.simibubi.create.Create;
+import com.simibubi.create.api.data.recipe.PolishingRecipeGen;
+
+import com.simibubi.create.api.data.recipe.SequencedAssemblyRecipeGen;
+
+import net.minecraft.data.PackOutput;
+
+@SuppressWarnings("unused")
+public final class CreatePolishingRecipeGen extends PolishingRecipeGen {
+
+	SequencedAssemblyRecipeGen.GeneratedRecipe
+
+	ROSE_QUARTZ = create(AllItems.ROSE_QUARTZ::get, b -> b.output(AllItems.POLISHED_ROSE_QUARTZ.get()))
+
+	;
+
+	public CreatePolishingRecipeGen(PackOutput output) {
+		super(output, Create.ID);
+	}
+}
