+++ src/main/java/com/simibubi/create/foundation/data/recipe/CreateSequencedAssemblyRecipeGen.java
@@ -1,23 +1,25 @@
 package com.simibubi.create.foundation.data.recipe;
 
-import java.util.function.UnaryOperator;
-import java.util.stream.Stream;
-
 import com.simibubi.create.AllBlocks;
 import com.simibubi.create.AllItems;
 import com.simibubi.create.AllTags.AllItemTags;
 import com.simibubi.create.Create;
+import com.simibubi.create.api.data.recipe.SequencedAssemblyRecipeGen;
+
 import com.simibubi.create.content.fluids.transfer.FillingRecipe;
 import com.simibubi.create.content.kinetics.deployer.DeployerApplicationRecipe;
 import com.simibubi.create.content.kinetics.press.PressingRecipe;
-import com.simibubi.create.content.processing.sequenced.SequencedAssemblyRecipeBuilder;
+
+import com.simibubi.create.foundation.data.recipe.CreateRecipeProvider.I;
 
 import net.minecraft.data.PackOutput;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.level.material.Fluids;
 
-public class SequencedAssemblyRecipeGen extends CreateRecipeProvider {
+import java.util.stream.Stream;
+
+public class CreateSequencedAssemblyRecipeGen extends SequencedAssemblyRecipeGen {
 
 	GeneratedRecipe
 
@@ -35,7 +37,7 @@ public class SequencedAssemblyRecipeGen extends CreateRecipeProvider {
 //		.addStep(FillingRecipe::new, rb -> rb.require(Fluids.WATER, 100))
 //		)
 
-	PRECISION_MECHANISM = create("precision_mechanism", b -> b.require(I.goldSheet())
+		PRECISION_MECHANISM = create("precision_mechanism", b -> b.require(I.goldSheet())
 		.transitionTo(AllItems.INCOMPLETE_PRECISION_MECHANISM.get())
 		.addOutput(AllItems.PRECISION_MECHANISM.get(), 120)
 		.addOutput(AllItems.GOLDEN_SHEET.get(), 8)
@@ -51,43 +53,29 @@ public class SequencedAssemblyRecipeGen extends CreateRecipeProvider {
 		.addStep(DeployerApplicationRecipe::new, rb -> rb.require(I.largeCog()))
 		.addStep(DeployerApplicationRecipe::new, rb -> rb.require(I.ironNugget()))),
 
-		REINFORCED_SHEET = create("sturdy_sheet", b -> b.require(AllItemTags.OBSIDIAN_DUST.tag)
-			.transitionTo(AllItems.INCOMPLETE_REINFORCED_SHEET.get())
-			.addOutput(AllItems.STURDY_SHEET.get(), 1)
-			.loops(1)
-			.addStep(FillingRecipe::new, rb -> rb.require(Fluids.LAVA, 500))
-			.addStep(PressingRecipe::new, rb -> rb)
-			.addStep(PressingRecipe::new, rb -> rb)),
+	REINFORCED_SHEET = create("sturdy_sheet", b -> b.require(AllItemTags.OBSIDIAN_DUST.tag)
+		.transitionTo(AllItems.INCOMPLETE_REINFORCED_SHEET.get())
+		.addOutput(AllItems.STURDY_SHEET.get(), 1)
+		.loops(1)
+		.addStep(FillingRecipe::new, rb -> rb.require(Fluids.LAVA, 500))
+		.addStep(PressingRecipe::new, rb -> rb)
+		.addStep(PressingRecipe::new, rb -> rb)),
 
-		TRACK = create("track", b -> b.require(AllItemTags.SLEEPERS.tag)
-			.transitionTo(AllItems.INCOMPLETE_TRACK.get())
-			.addOutput(AllBlocks.TRACK.get(), 1)
-			.loops(1)
-			.addStep(DeployerApplicationRecipe::new,
-				rb -> rb.require(Ingredient.fromValues(
-					Stream.of(new Ingredient.TagValue(I.ironNugget()), new Ingredient.TagValue(I.zincNugget())))))
-			.addStep(DeployerApplicationRecipe::new,
-				rb -> rb.require(Ingredient.fromValues(
-					Stream.of(new Ingredient.TagValue(I.ironNugget()), new Ingredient.TagValue(I.zincNugget())))))
-			.addStep(PressingRecipe::new, rb -> rb))
+	TRACK = create("track", b -> b.require(AllItemTags.SLEEPERS.tag)
+		.transitionTo(AllItems.INCOMPLETE_TRACK.get())
+		.addOutput(AllBlocks.TRACK.get(), 1)
+		.loops(1)
+		.addStep(DeployerApplicationRecipe::new,
+			rb -> rb.require(Ingredient.fromValues(
+				Stream.of(new Ingredient.TagValue(I.ironNugget()), new Ingredient.TagValue(I.zincNugget())))))
+		.addStep(DeployerApplicationRecipe::new,
+			rb -> rb.require(Ingredient.fromValues(
+				Stream.of(new Ingredient.TagValue(I.ironNugget()), new Ingredient.TagValue(I.zincNugget())))))
+		.addStep(PressingRecipe::new, rb -> rb))
 
-	;
+		;
 
-	public SequencedAssemblyRecipeGen(PackOutput p_i48262_1_) {
-		super(p_i48262_1_);
+	public CreateSequencedAssemblyRecipeGen(PackOutput output) {
+		super(output, Create.ID);
 	}
-
-	protected GeneratedRecipe create(String name, UnaryOperator<SequencedAssemblyRecipeBuilder> transform) {
-		GeneratedRecipe generatedRecipe =
-			c -> transform.apply(new SequencedAssemblyRecipeBuilder(Create.asResource(name)))
-				.build(c);
-		all.add(generatedRecipe);
-		return generatedRecipe;
-	}
-
-	@Override
-	public String getName() {
-		return "Create's Sequenced Assembly Recipes";
-	}
-
 }
