--- src/main/java/com/simibubi/create/foundation/data/recipe/DeployingRecipeGen.java
@@ -1,110 +1,0 @@
-package com.simibubi.create.foundation.data.recipe;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.function.Supplier;
-
-import com.simibubi.create.AllBlocks;
-import com.simibubi.create.AllRecipeTypes;
-import com.simibubi.create.foundation.block.CopperBlockSet;
-import com.simibubi.create.foundation.block.CopperBlockSet.Variant;
-
-import net.minecraft.data.PackOutput;
-import net.minecraft.tags.ItemTags;
-import net.minecraft.world.item.Items;
-import net.minecraft.world.level.ItemLike;
-import net.minecraft.world.level.block.Blocks;
-import net.minecraft.world.level.block.WeatheringCopper.WeatherState;
-
-public class DeployingRecipeGen extends ProcessingRecipeGen {
-
-	GeneratedRecipe COPPER_TILES = copperChain(AllBlocks.COPPER_TILES);
-	GeneratedRecipe COPPER_SHINGLES = copperChain(AllBlocks.COPPER_SHINGLES);
-
-	GeneratedRecipe
-
-	COGWHEEL = create("cogwheel", b -> b.require(I.shaft())
-		.require(I.planks())
-		.output(I.cog())),
-
-		LARGE_COGWHEEL = create("large_cogwheel", b -> b.require(I.cog())
-			.require(I.planks())
-			.output(I.largeCog()));
-
-	GeneratedRecipe
-
-	CB1 = addWax(() -> Blocks.WAXED_COPPER_BLOCK, () -> Blocks.COPPER_BLOCK),
-		CB2 = addWax(() -> Blocks.WAXED_EXPOSED_COPPER, () -> Blocks.EXPOSED_COPPER),
-		CB3 = addWax(() -> Blocks.WAXED_WEATHERED_COPPER, () -> Blocks.WEATHERED_COPPER),
-		CB4 = addWax(() -> Blocks.WAXED_OXIDIZED_COPPER, () -> Blocks.OXIDIZED_COPPER),
-
-		CCB1 = addWax(() -> Blocks.WAXED_CUT_COPPER, () -> Blocks.CUT_COPPER),
-		CCB2 = addWax(() -> Blocks.WAXED_EXPOSED_CUT_COPPER, () -> Blocks.EXPOSED_CUT_COPPER),
-		CCB3 = addWax(() -> Blocks.WAXED_WEATHERED_CUT_COPPER, () -> Blocks.WEATHERED_CUT_COPPER),
-		CCB4 = addWax(() -> Blocks.WAXED_OXIDIZED_CUT_COPPER, () -> Blocks.OXIDIZED_CUT_COPPER),
-
-		CCST1 = addWax(() -> Blocks.WAXED_CUT_COPPER_STAIRS, () -> Blocks.CUT_COPPER_STAIRS),
-		CCST2 = addWax(() -> Blocks.WAXED_EXPOSED_CUT_COPPER_STAIRS, () -> Blocks.EXPOSED_CUT_COPPER_STAIRS),
-		CCST3 = addWax(() -> Blocks.WAXED_WEATHERED_CUT_COPPER_STAIRS, () -> Blocks.WEATHERED_CUT_COPPER_STAIRS),
-		CCST4 = addWax(() -> Blocks.WAXED_OXIDIZED_CUT_COPPER_STAIRS, () -> Blocks.OXIDIZED_CUT_COPPER_STAIRS),
-
-		CCS1 = addWax(() -> Blocks.WAXED_CUT_COPPER_SLAB, () -> Blocks.CUT_COPPER_SLAB),
-		CCS2 = addWax(() -> Blocks.WAXED_EXPOSED_CUT_COPPER_SLAB, () -> Blocks.EXPOSED_CUT_COPPER_SLAB),
-		CCS3 = addWax(() -> Blocks.WAXED_WEATHERED_CUT_COPPER_SLAB, () -> Blocks.WEATHERED_CUT_COPPER_SLAB),
-		CCS4 = addWax(() -> Blocks.WAXED_OXIDIZED_CUT_COPPER_SLAB, () -> Blocks.OXIDIZED_CUT_COPPER_SLAB);
-
-	GeneratedRecipe
-	
-		CB_OX = oxidizationChain(List.of(() -> Blocks.COPPER_BLOCK, () -> Blocks.EXPOSED_COPPER, () -> Blocks.WEATHERED_COPPER, () -> Blocks.OXIDIZED_COPPER)),
-		CCB_OX = oxidizationChain(List.of(() -> Blocks.CUT_COPPER, () -> Blocks.EXPOSED_CUT_COPPER, () -> Blocks.WEATHERED_CUT_COPPER, () -> Blocks.OXIDIZED_CUT_COPPER)),
-		CCST_OX = oxidizationChain(List.of(() -> Blocks.CUT_COPPER_STAIRS, () -> Blocks.EXPOSED_CUT_COPPER_STAIRS, () -> Blocks.WEATHERED_CUT_COPPER_STAIRS, () -> Blocks.OXIDIZED_CUT_COPPER_STAIRS)),
-		CCS_OX = oxidizationChain(List.of(() -> Blocks.CUT_COPPER_SLAB, () -> Blocks.EXPOSED_CUT_COPPER_SLAB, () -> Blocks.WEATHERED_CUT_COPPER_SLAB, () -> Blocks.OXIDIZED_CUT_COPPER_SLAB));
-	
-	public GeneratedRecipe copperChain(CopperBlockSet set) {
-		for (Variant<?> variant : set.getVariants()) {
-			List<Supplier<ItemLike>> chain = new ArrayList<>(4);
-			
-			for (WeatherState state : WeatherState.values()) {
-				addWax(set.get(variant, state, true)::get, set.get(variant, state, false)::get);
-				chain.add(set.get(variant, state, false)::get);
-			}
-			
-			oxidizationChain(chain);
-		}
-		return null;
-	}
-
-	public GeneratedRecipe addWax(Supplier<ItemLike> waxed, Supplier<ItemLike> nonWaxed) {
-		createWithDeferredId(idWithSuffix(nonWaxed, "_from_removing_wax"), b -> b.require(waxed.get())
-			.require(ItemTags.AXES)
-			.toolNotConsumed()
-			.output(nonWaxed.get()));
-		
-		return createWithDeferredId(idWithSuffix(waxed, "_from_adding_wax"), b -> b.require(nonWaxed.get())
-			.require(Items.HONEYCOMB_BLOCK)
-			.toolNotConsumed()
-			.output(waxed.get()));
-	}
-	
-	public GeneratedRecipe oxidizationChain(List<Supplier<ItemLike>> chain) {
-		for (int i = 0; i < chain.size() - 1; i++) {
-			Supplier<ItemLike> to = chain.get(i);
-			Supplier<ItemLike> from = chain.get(i + 1);
-			createWithDeferredId(idWithSuffix(to, "_from_deoxidising"), b -> b.require(from.get())
-				.require(ItemTags.AXES)
-				.toolNotConsumed()
-				.output(to.get()));
-		}
-		return null;
-	}
-
-	public DeployingRecipeGen(PackOutput output) {
-		super(output);
-	}
-
-	@Override
-	protected AllRecipeTypes getRecipeType() {
-		return AllRecipeTypes.DEPLOYING;
-	}
-
-}
