+++ src/main/java/com/simibubi/create/infrastructure/data/CreateDatagen.java
@@ -11,10 +11,12 @@ import com.simibubi.create.AllSoundEvents;
 import com.simibubi.create.Create;
 import com.simibubi.create.foundation.advancement.AllAdvancements;
 import com.simibubi.create.foundation.data.DamageTypeTagGen;
-import com.simibubi.create.foundation.data.recipe.MechanicalCraftingRecipeGen;
-import com.simibubi.create.foundation.data.recipe.ProcessingRecipeGen;
-import com.simibubi.create.foundation.data.recipe.SequencedAssemblyRecipeGen;
-import com.simibubi.create.foundation.data.recipe.StandardRecipeGen;
+import com.simibubi.create.api.data.recipe.MechanicalCraftingRecipeGen;
+import com.simibubi.create.api.data.recipe.SequencedAssemblyRecipeGen;
+import com.simibubi.create.foundation.data.recipe.CreateMechanicalCraftingRecipeGen;
+import com.simibubi.create.foundation.data.recipe.CreateRecipeProvider;
+import com.simibubi.create.foundation.data.recipe.CreateSequencedAssemblyRecipeGen;
+import com.simibubi.create.foundation.data.recipe.CreateStandardRecipeGen;
 import com.simibubi.create.foundation.ponder.CreatePonderPlugin;
 import com.simibubi.create.foundation.utility.FilesHelper;
 import com.tterrag.registrate.providers.ProviderType;
@@ -47,13 +49,13 @@ public class CreateDatagen {
 		generator.addProvider(event.includeServer(), new CreateMountedItemStorageTypeTagsProvider(output, lookupProvider, existingFileHelper));
 		generator.addProvider(event.includeServer(), new DamageTypeTagGen(output, lookupProvider, existingFileHelper));
 		generator.addProvider(event.includeServer(), new AllAdvancements(output));
-		generator.addProvider(event.includeServer(), new StandardRecipeGen(output));
-		generator.addProvider(event.includeServer(), new MechanicalCraftingRecipeGen(output));
-		generator.addProvider(event.includeServer(), new SequencedAssemblyRecipeGen(output));
+		generator.addProvider(event.includeServer(), new CreateStandardRecipeGen(output));
+		generator.addProvider(event.includeServer(), new CreateMechanicalCraftingRecipeGen(output));
+		generator.addProvider(event.includeServer(), new CreateSequencedAssemblyRecipeGen(output));
 		generator.addProvider(event.includeServer(), new VanillaHatOffsetGenerator(output));
 
 		if (event.includeServer()) {
-			ProcessingRecipeGen.registerAll(generator, output);
+			CreateRecipeProvider.registerAllProcessing(generator, output);
 		}
 	}
